+++ src/test/java/pojo/storePojo/DeleteResponsePOJO.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigInteger;
+
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
 @JsonInclude(NON_NULL)
@@ -11,7 +13,7 @@ import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 @Builder
 public class DeleteResponsePOJO {
 
-    Integer code;
+    BigInteger code;
     String type;
     String message;
 }
