+++ src/test/java/pojo/storePojo/OrderPOJO.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigInteger;
 import java.util.Date;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
@@ -13,9 +14,9 @@ import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 @Builder
 public class OrderPOJO {
 
-    Integer id;
-    Integer petId;
-    Integer quantity;
+    BigInteger id;
+    BigInteger petId;
+    BigInteger quantity;
     Date shipDate;
     String status;
     Boolean complete;
