+++ src/test/java/pojo/userPojo/UserPOJO.java
@@ -1,9 +1,10 @@
 package pojo.userPojo;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+
+import java.math.BigInteger;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
@@ -12,7 +13,7 @@ import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 @JsonInclude(NON_NULL)
 public class UserPOJO {
 
-    Integer id;
+    BigInteger id;
     String username;
     String firstName;
     String lastName;
