+++ src/test/java/pojo/userPojo/UserResponseBodyPOJO.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigInteger;
+
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 
 @Getter
@@ -13,7 +15,7 @@ import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
 @Builder
 public class UserResponseBodyPOJO {
 
-    Integer code;
+    BigInteger code;
     String type;
     String message;
 }
